+++ NEWS.md
@@ -1,3 +1,7 @@
+# rasterpic (development version)
+
+-   Add support for `SpatVector`
+
 # rasterpic 0.1.0
 
 -   Initial release